Add `:legacy:ui:folder` module and move `FolderIconProvider`
