+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/messagedetails/MessageDetailsFragment.kt
@@ -20,6 +20,7 @@ import androidx.fragment.app.setFragmentResult
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.OnScrollListener
 import app.k9mail.core.ui.legacy.designsystem.atom.icon.Icons
+import app.k9mail.legacy.ui.folder.FolderIconProvider
 import app.k9mail.ui.utils.bottomsheet.ToolbarBottomSheetDialog
 import app.k9mail.ui.utils.bottomsheet.ToolbarBottomSheetDialogFragment
 import com.fsck.k9.activity.MessageCompose
@@ -29,7 +30,6 @@ import com.fsck.k9.mail.Address
 import com.fsck.k9.mailstore.CryptoResultAnnotation
 import com.fsck.k9.ui.R
 import com.fsck.k9.ui.base.extensions.withArguments
-import com.fsck.k9.ui.folders.FolderIconProvider
 import com.fsck.k9.ui.observe
 import com.mikepenz.fastadapter.FastAdapter
 import com.mikepenz.fastadapter.GenericItem
