+++ settings.gradle.kts
@@ -97,6 +97,7 @@ include(
     ":legacy:storage",
     ":legacy:testing",
     ":legacy:ui:base",
+    ":legacy:ui:folder",
     ":legacy:ui:legacy",
 )
